# docs/api/swagger.yaml
definitions:
...
  domain.Paginator:
    description: Paginator represents a paginator.
    properties:
      limit:
        description: Maximum number of items per page
        example: 10
        format: int
        type: integer
      next_page:
        description: URL for the next page
        example: http://localhost:8080/users?next_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      next_token:
        description: Base64 encoded next page token
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      prev_page:
        description: URL for the previous page
        example: http://localhost:8080/users?prev_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      prev_token:
        description: Base64 encoded previous page token
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      size:
        description: Number of items in the current page
        example: 10
        format: int
        type: integer
    type: object
...
  payload.HTTPMessage:
    description: HTTPMessage represents a message to be sent to the client trough
      HTTP REST API.
    properties:
      code:
        description: |-
          Code is a stable, machine-readable reason, present only where a client
          has to BRANCH on it. Message is prose and may be reworded; this is not.

          It exists because a 401 has two meanings a client must treat differently:
          an expired access token should be refreshed and the request retried,
          while a revoked one must not be — the refresh token was revoked in the
          same breath, so the retry burns two requests and fails anyway. Nothing in
          the old response distinguished them, and a client cannot be asked to
          match on English.

          Empty on every response that does not need one, so no existing client
          sees a change.
        example: token_revoked
        format: string
        type: string
      message:
        example: success
        format: string
        type: string
      method:
        example: GET
        format: string
        type: string
      path:
        example: /api/v1/users
        format: string
        type: string
      status_code:
        example: 200
        format: int32
        type: integer
      timestamp:
        example: "2021-07-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  payload.IDPTypesResponse:
    description: Complete identity provider type information including OAuth scopes
      and API endpoints
    properties:
      created_at:
        description: Timestamp when the identity provider type was created
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      description:
        description: Detailed description of the identity provider type and its capabilities
        example: Google OAuth2.0 Identity Provider
        format: string
        type: string
      id:
        description: Unique identifier for the identity provider type
        example: 019b4b0d-a682-7e30-8b33-650caa6446c7
        format: uuid
        type: string
      name:
        description: Display name of the identity provider type (e.g., Google, Github,
          Microsoft)
        example: Google
        format: string
        type: string
      scopes:
        description: OAuth scopes required for this identity provider type
        example:
        - openid
        - email
        - profile
        items:
          format: csv
          type: string
        type: array
      serial_id:
        description: Sequential identifier for ordering
        example: 1
        type: integer
      system:
        description: Indicates if this is a system-managed identity provider type
          (cannot be deleted)
        example: true
        type: boolean
      updated_at:
        description: Timestamp when the identity provider type was last updated
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      user_info_api_url:
        description: API endpoint URL to retrieve user information after authentication
        example: https://www.googleapis.com/oauth2/v3/userinfo
        format: uri
        type: string
    type: object
...
  payload.ListIDPTypesResponse:
    description: Paginated list of identity provider types available in the system
    properties:
      items:
        description: Array of identity provider type configurations
        items:
          $ref: '#/definitions/payload.IDPTypesResponse'
        type: array
      paginator:
        allOf:
        - $ref: '#/definitions/domain.Paginator'
        description: Pagination information including total count and page details
    type: object
...
info:
  contact:
    email: info@goapitemplate.local
    name: API Support
    url: https://goapitemplate.local
  description: |-
    A production-shaped starting point for a Go HTTP REST API: hexagonal architecture, multi-tenant projects, RBAC through Open Policy Agent, database-backed rate limiting and resource limits, and OpenTelemetry traces and metrics throughout.

    `products` is the worked example entity -- a project-scoped CRUD resource that exercises every convention here. Copy it when adding your own; see docs/architecture/adding-an-entity.md.

    ### Authentication

    Four bearer credentials, each accepted by a different set of endpoints; see the security definitions below for which is which. Ordinary calls use **AccessToken**.

    A logout revokes the access token it was called with, not only the refresh token, so it stops working immediately. That check can be switched off (`authn.access.token.revocation.enabled`), and where it is, a logged-out access token keeps working until it expires -- which is why the lifetime is short.

    Refresh tokens are rotated: each refresh spends the token it was given and returns a successor, so a client **must** store what `POST /auth/refresh` returns. Presenting a spent one again reads as a replay and ends the session for whoever holds the live token too.

    ### Conventions

    - **Identifiers are UUID v7.** Every `id` in this API, without exception.
...

    ### Rate limiting

...

...

...

    ### Operations

...
  version: v1
paths:
...
  /auth/idp_types:
    get:
      description: Retrieve a paginated list of all Identity Provider Types available
        for authentication.
      operationId: 0198f1e2-14ff-7678-afbe-9a627b0eaabd
      parameters:
      - description: Sort by fields (comma-separated with ASC/DESC)
        in: query
        name: sort
        type: string
      - description: Filter expression for querying results
        in: query
        name: filter
        type: string
      - description: Comma-separated list of fields to return
        in: query
        name: fields
        type: string
      - description: Pagination cursor for next page
        in: query
        name: next_token
        type: string
      - description: Pagination cursor for previous page
        in: query
        name: prev_token
        type: string
      - description: 'Maximum number of items to return (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of IDP types retrieved successfully
          schema:
            $ref: '#/definitions/payload.ListIDPTypesResponse'
        "400":
          description: Invalid request - malformed query parameters
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: List IDP types
      tags:
      - Authentication
      - Identity Provider Types
  /auth/idp_types/{idp_type_id}:
    get:
      description: Retrieve a specific Identity Provider Type by its unique identifier
      operationId: 0198f1e2-14ff-767c-971c-3904e0f2c484
      parameters:
      - description: IDP type unique identifier (UUID v7)
        in: path
        name: idp_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: IDP type details retrieved successfully
          schema:
            $ref: '#/definitions/payload.IDPTypesResponse'
        "400":
          description: Invalid request - malformed UUID or invalid parameters
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "404":
          description: IDP type not found
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: Get IDP type
      tags:
      - Authentication
      - Identity Provider Types
...
securityDefinitions:
  AccessToken:
    description: 'The ordinary credential: `Bearer <access_token>`, from `POST /auth/login`
      or `POST /auth/refresh`. Every endpoint outside `/auth`, `/health` and `/version`
      takes it. A personal access token is presented the same way and is accepted
      anywhere an access token is -- which is how a script authenticates without a
      login. Deliberately short-lived; refresh rather than lengthening it.'
    in: header
    name: Authorization
    type: apiKey
  RefreshToken:
    description: '`Bearer <refresh_token>`, and accepted **only** by `POST /auth/refresh`
      and `DELETE /auth/logout`. Refreshing spends it and returns a successor: store
      what comes back. Presenting a spent one is treated as a replay and ends the
      whole session, including for whoever holds the live token.'
    in: header
    name: Authorization
    type: apiKey
  ResetPasswordToken:
    description: '`Bearer <reset_password_token>`, accepted only by `POST /auth/password/reset`.
      Issued by `POST /auth/password/recover` and delivered by email, single use.'
    in: header
    name: Authorization
    type: apiKey
  VerificationToken:
    description: '`Bearer <verification_token>`, accepted only by `POST /auth/verify/confirm`.
      It proves an email address rather than exercising a permission, so it is the
      one credential no authorization check is applied to -- the account it belongs
      to is still disabled at that point. It never travels in a URL: the emailed link
      points at the frontend, which presents the token in this header.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...